+++ modules/tolyui_rx_layout/CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Support Row$ and Cell#span create responsive layout.
+
+## 0.0.2
+
+* Support Cell: [offset]、[push] 、[pull]
+* Support Row$: [verticalGutter]、[alignment]、[runAlignment]、[crossAxisAlignment]
+
+
+